fix: address round-4 review findings in capsule schema and API routes
- Change CapsuleMedia.sizeBytes from Int (INT32, overflows at 2 GB) to
  BigInt (BIGINT) in schema.prisma and migration.sql; serialize as string
  in API responses to avoid JS integer precision loss
- Add description length cap (10 000 chars) in POST /api/capsules and
  PATCH /api/capsules/[id] to prevent unbounded TEXT payloads
- Replace bare non-null assertion (updated!) in PATCH handler with an
  explicit null check returning 500, eliminating a potential runtime error
- Add comment to datasource db block explaining omission of url field
  (Prisma v5+ reads it from prisma.config.ts)

+++ prisma/migrations/20260514_init/migration.sql
@@ -35,7 +35,7 @@ CREATE TABLE "CapsuleMedia" (
     "url"       TEXT         NOT NULL,
     "filename"  TEXT         NOT NULL,
     "mimeType"  TEXT         NOT NULL,
-    "sizeBytes" INTEGER      NOT NULL,
+    "sizeBytes" BIGINT       NOT NULL,
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
 
     CONSTRAINT "CapsuleMedia_pkey" PRIMARY KEY ("id")
